WIP: Refactor to create app via a function, in preparation for creating a testing instance using a fake database.
